feat: update to rust 1.9.0

+++ src/sl.rs
@@ -4,16 +4,15 @@ pub mod logo;
 mod data;
 
 pub trait Train {
-  fn body(&self) -> &'static [&'static str];
-  fn wheelset(&self, x: uint) -> &'static [&'static str];
-  fn tender(&self) -> Option<&'static [&'static str]> {
-    None
-  }
-  fn wagons(&self) -> uint {
-    0
-  }
-  fn wagon(&self) -> Option<&'static [&'static str]> {
-    None
-  }
+    fn body(&self) -> &'static [&'static str];
+    fn wheelset(&self, x: usize) -> &'static [&'static str];
+    fn tender(&self) -> Option<&'static [&'static str]> {
+        None
+    }
+    fn wagons(&self) -> u32 {
+        0
+    }
+    fn wagon(&self) -> Option<&'static [&'static str]> {
+        None
+    }
 }
-
